requirements.txt
fastapi
uvicorn
python-dotenv
google-adk[web]
google-genai
PyPDF2
fpdf
python-docx
pydantic
